Changed Layer system, added support for inventory
Layer system split up in fore and background layers, foreground layers
can have entities, background cannot. Added the Inventory class and a
IInventory interface and made blocks able to be an inventory item. Also
removed the repeated Content.Load() in certain calls and changed it to a
single call to. Content.Load() at initialisation and storing the
reference to the texture in the class itself, this will come in usefull
when we want to change texture on the fly (with texturepacks for ex.)
Also renamed the Assembly and default namespace from "Cloud9" to
"SteamPilots".