Fixed airship, z-fighting player/airship, layer
Fixed airship shooting upwards at max speed when hitting a block. We
might need to redo the whole collision model later, but for now its
fixed. Also moved the Airship to layer 0 (foreground / grass  / etc)
layer so it doesn't z-fight with the player (that stayed on layer 1).
This might introduce the need for more layers (coulds / trees that you
want the airship to pass behind) but we can add that later on.