Added selection to the inventory top row.
Selection added, removed selection from the inventory background.
Need to clean the code a bit I think, messy with lots of referring etc
:S
Also the selection has to be repositioned and the scrolling won't works
as the value used to define how much slots to move is first cast into a
byte to be added to the slot which causes it to become 255.