Fixed gui backgrounds, tested scaling guis
Fixed gui backgrounds, when we switched towards positions instead of
rectangles the source rectangle was changed to be null if not told but
became 0 width and height. Fixed by setting the source by default and
give the choice to override.
Fixed a bug where you close the open guis with escape before going to
the pause menu, this caused the manager to close because it was only
used in the old gui system. Removed the code which causes the closing.
Not sure if it'll exist later on.
Tested around a bit with the scaling, to check it out see the inventory.
Need to find a good way to set positions etc correctly when changing
scale.